backend/requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.6.0
attrs==25.1.0
bcrypt==4.3.0
beanie==1.29.0
beautifulsoup4==4.13.3
blis==1.2.0
Brotli==1.1.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorclass==2.2.2
compressed_rtf==1.0.6
confection==0.1.5
cryptography==44.0.1
cssselect2==0.8.0
cymem==2.0.11
distro==1.9.0
dnspython==2.7.0
docx2txt==0.8
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
extract-msg==0.53.1
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lark==1.1.9
lazy-model==0.2.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.3
olefile==0.47
oletools==0.60.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pcodedmp==1.2.6
pdfminer.six==20240706
pillow==11.1.0
preshed==3.0.9
psutil==7.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
pyhtml2pdf==0.0.7
pymongo==4.11.1
pyparsing==3.2.1
PyPDF2==3.0.1
pyphen==0.17.2
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
fastapi-mail==1.4.1
APScheduler==3.10.4
python-jobspy @ git+https://github.com/jevil25/JobSpy.git
python-jose==3.4.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
red-black-tree-mod==1.22
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
resend==2.6.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
RTFDE==0.1.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.1
srsly==2.5.1
starlette==0.45.3
sympy==1.13.1
textract-py3==2.1.1
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tls-client==1.0.1
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.49.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
weasyprint==64.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.2
zopfli==0.2.3.post1
geoip2==4.8.0
